Feat: Update phone routing screen

+++ auto/src/test/java/com/framstag/libosmscout/client/FakeAutoRenderClient.kt
@@ -20,14 +20,14 @@ class FakeAutoRenderClient : OSMScoutClient() {
     override fun render(
         width: Int, height: Int,
         lat: Double, lon: Double,
-        angle: Double, magnification: Int
+        angle: Double, magnification: Double
     ): IntArray? {
         return IntArray(width * height) { 0xFFCCCCCC.toInt() }
     }
 
     override fun renderWithRouteAndPois(
         width: Int, height: Int,
-        lat: Double, lon: Double, angle: Double, magnification: Int,
+        lat: Double, lon: Double, angle: Double, magnification: Double,
         routeLats: DoubleArray?, routeLons: DoubleArray?,
         favoriteLats: DoubleArray?, favoriteLons: DoubleArray?,
         searchSelLat: Double, searchSelLon: Double,
